+++ lib/install/dotfiles/settings.sh
@@ -2,6 +2,11 @@
 # Restore ML4W Hyprland Settings app
 # ------------------------------------------------------
 
+echo -e "${GREEN}"
+figlet -f smslant "Settings"
+echo -e "${NONE}"
+
+
 if [ "$restored" == "1" ]; then
     if [ -f ~/.config/ml4w/settings/settings.json ] ;then
         # python restore.py
@@ -166,6 +171,53 @@ if [ -f ~/.config/ml4w/settings/waybar_window.sh ] ;then
     echo ":: waybar_window restored"
 fi
 
+# Replace waybar_settings
+if [ -f ~/.config/ml4w/settings/waybar_settings.sh ] ;then
+    replace_value=$(cat ~/.config/ml4w/settings/waybar_settings.sh)
+    search_str="settings"
+    if [[ $replace_value == "True" ]] ;then
+        replace_str="\"group\/settings\","
+    else
+        replace_str="\/\/\"group\/settings\","
+    fi
+    _replaceLineInFileCheckpoint $search_str $replace_str "modules-left" "$HOME/.config/waybar/themes/ml4w/config"
+    _replaceLineInFileCheckpoint $search_str $replace_str "modules-left" "$HOME/.config/waybar/themes/ml4w-blur/config"
+    _replaceLineInFileCheckpoint $search_str $replace_str "modules-left" "$HOME/.config/waybar/themes/ml4w-blur-bottom/config"
+    _replaceLineInFileCheckpoint $search_str $replace_str "modules-left" "$HOME/.config/waybar/themes/ml4w-bottom/config"
+    _replaceLineInFileCheckpoint $search_str $replace_str "modules-left" "$HOME/.config/waybar/themes/ml4w-minimal/config"
+    echo ":: waybar_settings restored"
+fi
+
+# Replace waybar_appmenu
+if [ -f ~/.config/ml4w/settings/waybar_appmenu.sh ] ;then
+    replace_value=$(cat ~/.config/ml4w/settings/waybar_appmenu.sh)
+    search_str="appmenu"
+    if [[ $replace_value == "True" ]] ;then
+        replace_str="\"custom\/appmenu\","
+    else
+        replace_str="\/\/\"custom\/appmenu\","
+    fi
+    _replaceLineInFile $search_str $replace_str "$HOME/.config/waybar/themes/ml4w/config"
+    _replaceLineInFile $search_str $replace_str "$HOME/.config/waybar/themes/ml4w-blur/config"
+    _replaceLineInFile $search_str $replace_str "$HOME/.config/waybar/themes/ml4w-blur-bottom/config"
+    _replaceLineInFile $search_str $replace_str "$HOME/.config/waybar/themes/ml4w-bottom/config"
+    _replaceLineInFile $search_str $replace_str "$HOME/.config/waybar/themes/starter/config"
+    echo ":: waybar_appmenu restored"
+fi
+
+# Replace waybar_appmenuicon
+if [ -f ~/.config/ml4w/settings/waybar_appmenu.sh ] ;then
+    replace_value=$(cat ~/.config/ml4w/settings/waybar_appmenu.sh)
+    search_str="appmenuicon"
+    if [[ $replace_value == "True" ]] ;then
+        replace_str="\"custom\/appmenuicon\","
+    else
+        replace_str="\/\/\"custom\/appmenuicon\","
+    fi
+    _replaceLineInFile $search_str $replace_str "$HOME/.config/waybar/themes/ml4w-minimal/config"
+    echo ":: waybar_appmenuicon restored"
+fi
+
 # Replace hypridle_hyprlock_timeout
 if [ -f ~/.config/ml4w/settings/hypridle_hyprlock_timeout.sh ] ;then
     replace_value=$(cat ~/.config/ml4w/settings/hypridle_hyprlock_timeout.sh)
@@ -192,3 +244,4 @@ if [ -f ~/.config/ml4w/settings/hypridle_suspend_timeout.sh ] ;then
     _replaceLineInFileCheckpoint "$search_str" "$replace_str" "SUSPEND TIMEOUT" "$HOME/.config/hypr/hypridle.conf"
     echo ":: hypridle_suspend_timeout restored"
 fi
+echo