+++ lib/install/packages/aur.sh
@@ -10,8 +10,8 @@ _installYay() {
     _installPackagesPacman "base-devel"
     SCRIPT=$(realpath "$0")
     temp_path=$(dirname "$SCRIPT")
-    git clone https://aur.archlinux.org/yay.git ~/yay
-    cd ~/yay
+    git clone https://aur.archlinux.org/yay.git ~/Downloads/yay
+    cd ~/Downloads/yay
     makepkg -si
     cd $temp_path
     echo ":: yay has been installed successfully."
@@ -21,8 +21,8 @@ _installParu() {
     _installPackagesPacman "base-devel"
     SCRIPT=$(realpath "$0")
     temp_path=$(dirname "$SCRIPT")
-    git clone https://aur.archlinux.org/paru.git ~/paru
-    cd ~/paru
+    git clone https://aur.archlinux.org/paru.git ~/Downloads/paru
+    cd ~/Downloads/paru
     makepkg -si
     cd $temp_path
     echo ":: paru has been installed successfully."
@@ -66,10 +66,9 @@ _checkAURHelper() {
     fi
 }
 
-if [[ $(_check_update) == "false" ]] ;then
-    echo -e "${GREEN}"
-    figlet -f smslant "AUR Helper"
-    echo -e "${NONE}"
-fi
+echo -e "${GREEN}"
+figlet -f smslant "AUR Helper"
+echo -e "${NONE}"
+
 _checkAURHelper