+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+Version 2.9.6.4
+--------------------------------------------------------
+- Kitty is the new default terminal (Replace alacritty with kitty in the ML4W Settings App/Default terminal)
+- Fixed eza icons on zsh terminal 
+- Subfolders in .config can be protected with an empty file PROTECTED
+
+Version 2.9.6.3
+--------------------------------------------------------
+- Updated ags sidebar layout
+- Calendar widget shows always current date when opening
+- Wallpaper, Effects and waybar theme switcher button moved into sidebar
+- Settings group in waybar can be hidden in the ML4W Settings App
+- Wallpaper cache covering also blurred images
+- Added "Open in terminal" feature to Nautilus
+- New default wallpaper.
+
+Version 2.9.6.2
+--------------------------------------------------------
+- App Menu icon can be hidden from the ML4W Settings App
+- Setup script optimize the keybindings if french keyboard layout (fr) is selected
+- Add full waypaper features with waypaper 2.3. Brings back swww support including animations.
+- ML4W apps will be now installed into the linux file system 
+
 Version 2.9.6.1
 --------------------------------------------------------
 - AGS calendar widget moved from sidebar into own widget. Opens with click on clock module in waybar